remove hashmaps, use only function execution info on function rel score

+++ common/src/config.rs
@@ -46,7 +46,7 @@ pub const MAX_SPLICE_TIMES: usize = 45;
 
 
 //Cheong
-pub const FUNC_REL_HIGH_THRESHOLD : f64 = 0.7;
+pub const FUNC_REL_HIGH_THRESHOLD : f32 = 0.7;
 pub const FUZZ_TIME_OUT : u32 = 60 * 60 * 24;
 
 //MUTATE TC SELECTION assume either one of them or none of them
